refactor: Move game related things to the game module
